+++ lib/rebuilder.dart
@@ -4,4 +4,4 @@ export 'src/data_model.dart';
 export 'src/data_provider.dart';
 export 'src/rebuilder.dart';
 export 'src/rebuilder_object.dart';
-export 'src/state_wrapper.dart';
+export 'src/rebuilder_state.dart';