+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 0.2.0 (08-06-19)
+
+- *Breaking change*: the class `StateWrapper` was renamed to `RebuilderState`.
+- Code refactoring
+- Updated readme.md: added a trivia game example.
+
 ## Version 0.1.0+1 (31-05-19)
 
 - First version of the package